==> 9. Transforming data/2_Pair_Prog_Limpieza_2.ipynb <==
# 04/05 - Pair Programming Limpieza 1 - Guadalupe & Lara

import pandas as pd
import seaborn as sns
import numpy as np
import re
import matplotlib.pyplot as plt
pd.options.display.max_columns = None
df = pd.read_csv("1.ataques.csv", index_col = 0)
df.head()

# EJERCICIOS

# 1 Columna de species:

Si exploramos esta columna en detalle nos podemos dar cuenta que tenemos muchos valores únicos y esto hace que sea muy difícil trabajar con esta columna. Lo que vamos a hacer es clasificar los tiburones en 5 especies diferentes, las más comunes, que incluyen el tiburón blanco (White), el tiburón tigre (Tiger ), el tiburón gris (Grey), el tiburón limón (Lemon) y el tiburón toro (Bull). 
El resto de las especies las incluiremos en un único grupo que podremos llamar "Unspecified". Ahora nos podemos sentir un poco abrumadas y no saber como enfrentarnos a este reto, pero don't worry, os dejamos por aquí unas 💡 pistas 💡 para que os ayuden a entender cómo hacerlo.
Los valores de las columnas son strings por lo que podremos usar regex para buscar palabras clave en cada celda y asignarlo a una de las categorías que hemos definido previamente.

 *imaginemos que el valor de una celda es el siguiente*

*'White shark, 3.5 m'*

* tendremos que buscar el patrón de regex que nos permita extraer White shark de ese string y que nos devuelva solo White shark. *
* Un patrón que podriamos usar es: *

*patron_blanco = r".*[Ww](hite|HITE).*" # esto podría ser así porque puede estar en mayúsculas o en minúsculas. *

* de la misma forma que hemos sacado el patron para el tiburón blanco, tendremos que sacar los patrones para las otras 4 especies que queremos "encontrar". *

Tendremos que crearnos una función que aplicaremos sobre nuestra columna species_ para que nos devuelva una nueva columna con los valores clasificados en función de los patrones de regex que hayamos definido.

patron_blanco =r".*[Ww](hite|HITE).*"
patron_tigre =r".*[Tt](iger|IGER).*"
patron_gris = r".*[Gg](rey|REY).*"
patron_limon =r".*[Ll](emon|EMON).*"
patron_torito = r".*[Bb](ull|ULL).*"


def funcion (columna):
    
    try:
        if re.findall(patron_blanco,columna) :
            return "white_shark"
        elif re.findall(patron_tigre,columna) :
            return "tiger_shark"
        elif re.findall(patron_gris,columna) :
            return "grey_shark"
        elif re.findall(patron_limon,columna) :
            return "lemon_shark"
        elif re.findall(patron_torito,columna) :
            return "bull_shark"
        else:
            return "unspecified"
    except:
        return "unspecified"

df["species"] = df["species_"].apply(funcion)
df[["species_", "species"]].head(10)

df["species"].value_counts()

# EJERCICIO 2: Columna "age"

patron_age = "[0-9]{1,2}" 
df['age'] = df['age'].astype(str)

def limpio(col):
    x = re.findall(patron_age, col)
    try:
        return int(x[0])
    except:
        np.NaN

df["age2"] = df["age"].apply(limpio)

df["age2"].unique()

df['age2'].dtypes

*Una vez que hayáis extraido los números, os daréis cuenta que hay celdas que tienen más de una edad. Tendréis que decidir que hacer en esos casos. Os dejamos por aquí una posible opción usando un método de Pandas que os puede resultar super útil. El **método explode**, aqui.*

-Por último cambiad el tipo de la columna de string a integer.

df["age2"] = df["age2"].astype("Int64", errors = "ignore")
df.dtypes

# 3
Guardad el csv con las columnas limpias para seguir trabajando con este dataframe limpio.

df.to_csv("2.ataques.csv")
