==> 9. Transforming data/5_Pair_Prog_Limpieza_5.ipynb <==
# 11/05 - Pair Programming Limpieza 5 - Guadalupe & Lara

import pandas as pd
import seaborn as sns
import matplotlib.pyplot as plt
import numpy as np
from sklearn.impute import SimpleImputer
from sklearn.experimental import enable_iterative_imputer
from sklearn.impute import IterativeImputer
from sklearn.impute import KNNImputer
df = pd.read_csv("4.Ataques_enproceso_4.csv", index_col = 0)

1- **Es el momento de eliminar los nulos:**

-*Reemplazad los valores nulos de la columna age por la media de la edad usando el método SimpleImputer.*


imputer = SimpleImputer(strategy='mean', missing_values=np.nan)

imputer = imputer.fit(df[['age2']])

df['age_mean'] = imputer.transform(df[['age2']])

df.isnull().sum()

-*Reemplazad los valores nulos de la columna sex por la moda, usando el método SimpleImputer.*

💡 ** Pista** 💡 La moda en este tipo de aproximación se indica como most_frequent.


imputer2 = SimpleImputer(strategy='most_frequent', missing_values=np.nan)

imputer2 = imputer2.fit(df[['sex_']])

df['sex_moda'] = imputer2.transform(df[['sex_']])

df['sex_'].value_counts()

df['sex_moda'].value_counts()

Reemplazad los valores nulos de la columna type por el valor más frecuente (la moda) con el método SimpleImputer.*


imputer3 = imputer2.fit(df[['type']])
df['type_moda'] = imputer3.transform(df[['type']])

-*Utilizad el método KNN Imputer para reemplazar todos los valores nulos de las columnas numéricas.*

-*Utilizad el método Iterative Imputer para reemplazar todos los valores nulos de las columnas numéricas.*

-*¿Podríais explicar que diferencia hay entre estos dos últimos métodos?*

Guardad el csv para seguir trabajando con el en los siguientes ejercicios de pair

numericas = df.select_dtypes(include = np.number)
numericas.isnull().sum()

imputerKNN = KNNImputer(n_neighbors=5)
imputerKNN.fit(numericas)
numericas_knn= imputerKNN.transform(numericas)

df_knn_imputer = pd.DataFrame(numericas_knn)
df_knn_imputer.head()

# df_knn_imputer = pd.DataFrame(numericas_knn, columns = ['yearknn', 'age2knn', 'age2fillnaknn', 'age_meanknn']) Solo nos coge 3 columnas, no 4, por eso da error.
df_knn_imputer = pd.DataFrame(numericas_knn, columns = ['yearknn', 'age2knn', 'age_meanknn'])
df_knn_imputer.head()

Guardad el csv para seguir trabajando con el en los siguientes ejercicios de pair

df.to_csv("5.Ataques_enproceso_5.csv")
